Copy upload-to-s3 from ncov-ingest
Copied from https://github.com/nextstrain/ncov-ingest/blob/6fd5a9b1d87e59fab35173dbedf376632154943b/bin/upload-to-s3

Copies of this script, from an earlier version¹, are missing zstd
compression support:
- https://github.com/nextstrain/rsv/blob/ba171f4a43110382c38b6154be3febd50408d7bf/ingest/bin/upload-to-s3
- https://github.com/nextstrain/dengue/blob/247b2fd897361f2548627de1d97d45fae4115c5c/ingest/bin/upload-to-s3
- https://github.com/nextstrain/zika/blob/7ec737b8d249302e9dc1e26f42e32696781e1488/ingest/bin/upload-to-s3

Copies of this script that are functionally identical:
- https://github.com/nextstrain/monkeypox/blob/5c461dc7e90cd70c1f16b193f82fd1666d4c95e2/ingest/bin/upload-to-s3
- https://github.com/nextstrain/forecasts-ncov/blob/fcc7c881188f6d8958cde17a62fd7483738c6a36/ingest/bin/upload-to-s3

The copy of this script in ncov² has diverged enough that I think it's
not worth it to incorporate changes here. Although, a future improvement
that would be nice to have is a way to bypass the compression during
upload if the local file is already compressed.

¹ https://github.com/nextstrain/ncov-ingest/blob/5424b34595b46cc7295262ea7084feccbb67ae2d/bin/upload-to-s3
² https://github.com/nextstrain/ncov/blob/ba9953f8a61d239176e960b9b2efb925b5dff84b/scripts/upload-to-s3