download-from-s3: Add comment on subsample option
Add a comment to note that this is intended to be used for debugging
workflows only and should not be used for actual subsampling.

Also, I totally missed in my review of https://github.com/nextstrain/ncov-ingest/pull/367
that this option will be used for FASTA files, which results in a bad
mixture of sequence IDs and sequences. This is ~fine for ncov-ingest
because it is only being used for debugging the workflow, but definitely
should not be used for production workflows.