notify-on-record-change: generalize messages
Generalize the error messages for the `src` and `dst` arguments by
removing mentions of "GISAID" and "GenBank" and generalize the Slack
messages by removing "nCoV".

Instead of injecting pathogen names with a new argument, I think it will
be sufficient to provide the custom message through the third `source_name`
argument.

Based on diffs with subsequent copies of the script that edited them:
- https://github.com/nextstrain/monkeypox/blob/5c461dc7e90cd70c1f16b193f82fd1666d4c95e2/ingest/bin/notify-on-record-change
- https://github.com/nextstrain/rsv/blob/ba171f4a43110382c38b6154be3febd50408d7bf/ingest/bin/notify-on-record-change
- https://github.com/nextstrain/dengue/blob/247b2fd897361f2548627de1d97d45fae4115c5c/ingest/bin/notify-on-record-change
- https://github.com/nextstrain/zika/blob/4ac8d526f9f14be10b7e8858ad469a40b72a505e/ingest/bin/notify-on-record-change