fetch-from-ncbi-entrez: Use the Entrez History server
Use the Entrez History server to store search results rather than iterate
over batches of UIDs. This allows us to avoid round-tripping the
UIDs returned by the ESearch so that we are no longer limited by the
200 UID limit for EFetch.

Also removing the hardcoded `MAX_IDS` limit set for HepB since we can
get the total number of records from the ESearch results.