ncbi-virus-url: Add `updated` as an output column
Useful for knowing when the record was last updated in NCBI, which
@j23414 noted to be the "Modification date".¹

This will ultimately add a new field to the NDJSON output from
fetch-from-ncbi-virus, but it should not affect the final metadata
output of the ingest pipelines. Ingest pipelines usually define the
final metadata output columns in the config and would need to be updated
separately.

¹ https://github.com/nextstrain/monkeypox/issues/104#issuecomment-1320342545