Use the environment's Bash
We definitely don't need "the system bash", which is what /bin/bash is.
We want "the bash the user wants", which will often be the system bash,
but not always.

The main motivation is that /bin/bash on macOS is stuck on an ancient
version due to licensing issues, and users might prefer another Bash in
their environments.