Remove NCBI Virus API related files
Due to https://github.com/nextstrain/ingest/issues/18, the NCBI Virus
API is more of a hassle to use. The data from NCBI Datasets CLI covers
the standard fields that we use in pathogen ingests, so we can drop
the use of the undocumented NCBI Virus API.

If any pathogen needs additional custom fields that are not available
through NCBI Datasets, the pipeline can use fetch-from-ncbi-entrez
and parse the GenBank file.