README: Document workaround with subrepo's `--force` flag
In walking through `git subrepo pull` with @j23414 for monkeypox, we
ran into unexpected merge conflicts.¹

Reading through the git subrepo docs,² I found that we can just make a
clean "reclone" of the subrepo with the `--force` flag. Instead of
working through the conflict resolution, the forced reclone made sense
for just getting the latest version of the ingest repo.

¹ https://gist.github.com/joverlee521/b00739c3dcbff66ca99828d43cb80a42
² https://github.com/ingydotnet/git-subrepo#commands