Update `transform-genbank-location` to complain... [#35]
...about violated expectations around `database` and `location` fields
in the input record.

Note: I choose to have the `parse_location` method bail early and
return an unmodified record if the `location` field isn't found. The
other alternative is to issue the warning but add the empty `country`,
`division`, and `location` records. I don't have strong feelings about
this one way or the other; if somebody does, I'll change this
behavior.