notify-slack: Migrate to new upload APIs
The `files.upload` API will be sunset on March 11, 2025.¹ This commit
follows Slack's "Uploading files" docs² to migrate to the new upload
APIs, `files.getUploadURLExternal` and `files.completeUploadExternal`.

The usage of the script does not change. However, the new API
requires the channels to be provided as their channel IDs instead of
channel names, so we will have to update the `SLACK_CHANNELS` envvar
for various pathogen workflows. Channel ids are more stable anyways
since channels can be renamed.

¹ <https://api.slack.com/changelog/2024-04-a-better-way-to-upload-files-is-here-to-stay>
² <https://api.slack.com/messaging/files#uploading_files>