Add shared snakemake directory
Adds the shared config parsing function in config.smk

This file is expected to hold functions for parsing Snakemake workflow configs. 
It currently only includes the `resolve_config_path` function that is needed to 
support external analysis directories with `nextstrain run`. 

The `resolve_config_path` function is built on the work @tsibley did in 
measles and the work @jameshadfield did in avian-flu.

Note the optional argument *other_prefixes* is added to mainly to support 
avian-flu's directory structure that does not match the pathogen-repo-guide. 
This will allow avian-flu to pass in custom prefixes to check for config paths.

<https://github.com/nextstrain/measles/blob/efda4428d83bf0de092db21d2dd00822a0a41509/shared/functions.smk#L3>
<https://github.com/nextstrain/avian-flu/blob/71f005f94f4256e959c90a2830c646a660137a75/rules/config.smk#L26>